+++ lib/src/core/models/requests/mobile_money/mobile_money_request_body.dart
@@ -10,7 +10,7 @@ class MobileMoneyRequestBody {
   });
 
   String? txRef;
-  int? amount;
+  double? amount;
   String? email;
   String? phoneNumber;
   String? currency;
